added type conversion

+++ include/memory.hpp
@@ -20,14 +20,14 @@ public:
     std::size_t inst_size() { return mem.size(); };
 
     void write(int32_t effective_address, int32_t value) {
-        if ((effective_address < 0) || (effective_address >= mem.capacity()))
+        if ((effective_address < 0) || (effective_address >= (int32_t)mem.capacity()))
             throw std::runtime_error("Cannot write to memory, invalid address");
 
         mem[effective_address] = value;
     }
 
     uint32_t read(int32_t effective_address) const {
-        if ((effective_address < 0) || (effective_address >= mem.size()))
+        if ((effective_address < 0) || (effective_address >= (int32_t)mem.size()))
             throw std::runtime_error("Cannot read from memory, invalid address");
 
         return mem[effective_address];
